+++ xknx/remote_value/remote_value_temp.py
@@ -6,24 +6,17 @@ DPT 9.001.
 from __future__ import annotations
 
 from xknx.dpt import DPTArray, DPTBinary, DPTTemperature
-from xknx.exceptions import CouldNotParseTelegram
 
 from .remote_value import RemoteValue
 
 
-class RemoteValueTemp(RemoteValue[DPTArray, float]):
+class RemoteValueTemp(RemoteValue[float]):
     """Abstraction for remote value of KNX 9.001 (DPT_Value_Temp)."""
-
-    def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTArray:
-        """Test if telegram payload may be parsed."""
-        if isinstance(payload, DPTArray) and len(payload.value) == 2:
-            return payload
-        raise CouldNotParseTelegram("Payload invalid", payload=str(payload))
 
     def to_knx(self, value: float) -> DPTArray:
         """Convert value to payload."""
         return DPTTemperature.to_knx(value)
 
-    def from_knx(self, payload: DPTArray) -> float:
+    def from_knx(self, payload: DPTArray | DPTBinary) -> float:
         """Convert current payload to value."""
-        return DPTTemperature.from_knx(payload.value)
+        return DPTTemperature.from_knx(payload)